[WCNCR00262566] wlan: refactor pointer casting
[Description]
refactor uintptr_t as unsigned long for different OS pointer type on x64/32
add define of uintptr_t if not in other OS

MTK-Commit-Id: da16bd5f8f9e49e388ed58b3805a97c3eeb4e398

CR-Id: WCNCR00262566
Feature: os/wlan
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6073805
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>

+++ common/dump.c
@@ -363,9 +363,9 @@ void dumpMemory32(IN uint32_t *pu4StartAddr,
 	LOG_FUNC("DUMP32 ADDRESS: 0x%x, Length: %d\n", pu4StartAddr,
 		 u4Length);
 
-	if (IS_NOT_ALIGN_4((unsigned long)pu4StartAddr)) {
+	if (IS_NOT_ALIGN_4((uintptr_t)pu4StartAddr)) {
 		uint32_t u4ProtrudeLen =
-			sizeof(uint32_t) - ((unsigned long)pu4StartAddr % 4);
+			sizeof(uint32_t) - ((uintptr_t)pu4StartAddr % 4);
 
 		u4ProtrudeLen =
 			((u4Length < u4ProtrudeLen) ? u4Length : u4ProtrudeLen);
@@ -391,7 +391,7 @@ void dumpMemory32(IN uint32_t *pu4StartAddr,
 
 		u4Length -= u4ProtrudeLen;
 		pu4StartAddr = (uint32_t *)
-			       ((unsigned long)pu4StartAddr + u4ProtrudeLen);
+			       ((uintptr_t)pu4StartAddr + u4ProtrudeLen);
 	}
 
 	while (u4Length > 0) {
@@ -529,9 +529,9 @@ void dumpMemory128(IN uint32_t *pu4StartAddr,
 	LOG_FUNC("DUMP32 ADDRESS: 0x%x, Length: %d ", pu4StartAddr,
 		 u4Length);
 
-	if (IS_NOT_ALIGN_4((unsigned long)pu4StartAddr)) {
+	if (IS_NOT_ALIGN_4((uintptr_t)pu4StartAddr)) {
 		uint32_t u4ProtrudeLen =
-			sizeof(uint32_t) - ((unsigned long)pu4StartAddr % 4);
+			sizeof(uint32_t) - ((uintptr_t)pu4StartAddr % 4);
 
 		u4ProtrudeLen =
 			((u4Length < u4ProtrudeLen) ? u4Length : u4ProtrudeLen);
@@ -557,7 +557,7 @@ void dumpMemory128(IN uint32_t *pu4StartAddr,
 
 		u4Length -= u4ProtrudeLen;
 		pu4StartAddr = (uint32_t *)
-			       ((unsigned long)pu4StartAddr + u4ProtrudeLen);
+			       ((uintptr_t)pu4StartAddr + u4ProtrudeLen);
 	}
 
 	while (u4Length > 0) {
